+++ triple_arena/CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## [0.6.0] - TODO
 ### Changes
+- Major overhaul of `Ptr`s that reduces `Ptr<P0>` down to just `P0` and allows for different index
+  and generation number sizes
 - The `PartialOrd` on `Ptr`s now sorts first by the internal raw index followed by generation value
 
 ## [0.5.0] - 2022-05-31
 ### Additions
-- Added `Default` bounds and impls for `PtrTrait` and `Ptr<P>`
+- Added `Default` bounds and impls for `Ptr` and `P`
 
 ## [0.4.1] - 2022-04-09
 ### Fixes
@@ -16,7 +18,7 @@
 
 ## [0.4.0] - 2022-01-17
 ### Changes
-- Added `Send` and `Sync` bounds to `PtrTrait`
+- Added `Send` and `Sync` bounds to `Ptr`
 - Added an argument to `create` in `Arena::try_insert_with`
 - The iteration structs are now public